Rename UserContext and refactor to store all auth state and methods in context file and provide through custom context component
